Conflicts:
	master.c

+++ acoustic_modem.c
@@ -5,14 +5,15 @@
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
-#define BUFSIZE 256/*default size for buffer*/
+#include <time.h>
+#define BUFSIZE 100/*default size for buffer*/
 #define SERIAL_TIMEOUT 2000/*default timeout for reading modem*/
-#define TX_LOG "echo $(date -Ru) >> /home/root/log/TXLOG.TXT"
-#define TX_PATH "/home/root/log/TXLOG.TXT"
-#define RX_LOG "echo $(date -Ru) >> /home/root/log/RXLOG.TXT"
-#define RX_PATH "/home/root/log/RXLOG.TXT"
 #define AMODEM_PATH "/home/root/log/AMODEM.TXT"
-#define GPSPIPE "gpspipe -r -n 12 |grep 'GPGGA' >> /dev/ttyUSB2" /*feed modem gps GPGGA setence.*/
+#define GPSPIPE "gpspipe -r -n 20 |grep 'GPGGA' >> /dev/ttyUSB2" /*feed modem gps GPGGA setence.*/
+#define TX_PATH "/home/root/log/TXLOG.TXT"
+#define RX_PATH "/home/root/log/RXLOG.TXT"
+#define SYCN_TIMEOUT 15
+#define COPY_TIMEOUT 60000
 
 a_modem modem;/*a struct that contains the status of modem or some useful information*/
 a_modem_msg msg;/*a list that contains latest msg from (local) modem*/
@@ -20,36 +21,40 @@ a_modem_msg msg_remote;/*a list that contains latest msg from (remote) modem*/
 
 
 int a_modem_ffs_clear(){
+/*clear ffs system, high usage ie. 98% of ffs system make the modem
+behave strange
+rm *.log file and *.wav file except *t*.wav file
+*/
 char fname[80];
 int rm;
 a_modem_puts("ls /ffs/ \r");
-while (a_modem_wait_info(NULL,SERIAL_TIMEOUT,fname,80)>1){
+while (a_modem_wait_info(NULL,SERIAL_TIMEOUT,fname,80)==SUCCESS){
 printf("remove file : %s\n",fname);
 rm=0;
+if ((strstr(fname,"user")!=NULL)&&(strlen(fname)<7))continue;
 if (strstr(fname,"log")!=NULL) rm=1;
 if ((strstr(fname,"wav")!=NULL)&&(strstr(fname,"t")==NULL)) rm=1;
 if (rm){
 a_modem_puts("rm /ffs/");
 a_modem_puts(fname);
-a_modem_puts("\r");
+a_modem_puts("\r");}
+usleep(100000);
+fname[0]=0;
 }
-}
-
-
 return SUCCESS;
 }
-void a_modem_print(int timeout){
+
+void a_modem_print(int msec){
 /*print all the text from serial port*/
-int Niter=timeout/N_ITER_DIV,delay=0;
+int Niter=msec/N_ITER_DIV,delay=0;
 char buf[BUFSIZE];
 
 while(delay<Niter){
-if (a_modem_gets(buf,BUFSIZE)==FAIL)delay++;
-else printf("%s\n",buf);
+if (a_modem_gets(buf,BUFSIZE)==FAIL){
+delay++;
 usleep(WAIT_INTVAL);
+}else printf("%s\n",buf);
 }
-
-return SUCCESS;
 }
 
 int a_modem_msg_add(a_modem_msg *msg_list ,char *msg_str){
@@ -66,19 +71,18 @@ int a_modem_msg_add(a_modem_msg *msg_list ,char *msg_str){
 	return SUCCESS;
 }
 
-int a_modem_wait_remote(char *buf,int bufsize,int timeout){
+int a_modem_wait_remote(char *buf,int bufsize,int msec){
 /*read msg from remote list and store in buf until timeout (miliseconds)*/
-int delay=0,Niter=timeout/N_ITER_DIV;
+int delay=0,Niter=msec/N_ITER_DIV;
 
 while(delay<Niter){//before timeout
 if (msg_remote.N_unread>0){/*read the oldest msg from remote msg list*/
 if (buf!=NULL)strcpy(buf,MSG_PULL(msg_remote));
 msg_remote.N_unread--;
 return SUCCESS;
-break;
 }
 
-a_modem_gets(NULL,BUFSIZE);
+a_modem_gets(NULL,0);
 usleep(WAIT_INTVAL);
 delay++;
 }
@@ -98,8 +102,9 @@ void a_modem_msg_show(a_modem_msg * list){
 }
 
 int a_modem_init(){
-	/*init the struct variable*/
 	int i;
+
+	/*init the struct variable*/
 	/*init msg list (remote & local)*/
 	msg.i=0;
 	msg.N_unread=0;
@@ -110,13 +115,7 @@ int a_modem_init(){
 	msg.text[i]=strdup(" ");
 	msg_remote.text[i]=strdup("");
 	}
-	/*init modem struct*/
-	/*if (t_node.this_node->tx_fname[0]==0){
-	printf("fail to find default tx wavform name\n");
-	}else{
-	printf("find default tx wav filename\n");
-	strcpy(modem.def_tx_wav,t_node.this_node->tx_fname);
-	}*/
+	/*init modem*/
 	modem.latest_tx_stamp[0]=0;
 	modem.latest_rx_fname[0]=0;
 	modem.def_tx_wav[0]=0;
@@ -125,6 +124,15 @@ int a_modem_init(){
 	modem.dsp_bat=0;
 	modem.mdm_bat=0;
 	modem.rtc_bat=0;
+	/*TX / RX log*/
+	if ((modem.tx_p=fopen(TX_PATH,"a"))==NULL){
+	fprintf(stderr,"fail to open TX log\n");
+	return FAIL;
+	}
+	if ((modem.rx_p=fopen(RX_PATH,"a"))==NULL){
+	fprintf(stderr,"fail to open RX log\n");
+	return FAIL;
+	}
 	return SUCCESS;
 }
 
@@ -155,7 +163,7 @@ int a_modem_cfg_set(const char *fname) {
 	char buf[BUFSIZE];
 	a_modem_clear_io_buffer();
 	a_modem_puts("+++\r");
-	if (fp=fopen(fname,"r")==NULL){
+	if ((fp=fopen(fname,"r"))==NULL){
 	fprintf(stderr,"fail to open %s\n",CFG_DEVEL);
 	return FAIL;
 	}
@@ -180,10 +188,10 @@ int a_modem_cfg_set(const char *fname) {
 int a_modem_play(char * filename) {
 	//play a wavform, store the tx time
 	char buf[BUFSIZE];
-	char buf2[BUFSIZE];
-	int n;
 	modem.latest_tx_stamp[0]=0;
-// play
+	time_t bb_stamp;
+	time(&bb_stamp);
+	// play
 	a_modem_puts("\r");
 	sprintf(buf, "play /ffs/%s\r", filename); 
 	if (a_modem_puts(buf)==FAIL) {
@@ -194,14 +202,14 @@ int a_modem_play(char * filename) {
 		fprintf(stderr, "A_modem, fail to play waveform\n");
 		return FAIL;
 	}
-// get time stamp
+	// get time stamp
 	//TODO buffer time is random
 	if (a_modem_wait_info("tx",WAIT_TXTIME, buf, BUFSIZE)==SUCCESS) {
-		printf("info : %s\n", buf);
+		printf("local tx time : %s\n", buf);
 		strcpy(modem.latest_tx_stamp,buf);
-		sprintf(buf2, "echo '%s,%s' >> %s", filename, buf,TX_PATH);
-		system(buf2); //TODO store result using graceful way
-		system(TX_LOG);
+		fprintf(modem.tx_p,"%s,%s\n",filename,buf);
+		fprintf(modem.tx_p,"BB:%s\n\n",ctime(&bb_stamp));
+		fflush(modem.tx_p);
 		return SUCCESS;
 	} else {
 		printf("A_modem, fail to get TX time\n");
@@ -216,7 +224,7 @@ inline int a_modem_puts(const char*msg){
 
 int a_modem_gets(char* buf,int size){
 	/* read a line of text from serial port
-	Input 
+	Input
 		*buf can be char array ready for store the string 
 		or NULL
 		*size the array size of buf
@@ -246,8 +254,8 @@ int a_modem_gets(char* buf,int size){
 	dump[n-1]=0;/*remove newline char*/
 	if (strstr(dump,"DATA")==NULL)	a_modem_msg_add(&msg,dump);
 	else{
-	sscanf(dump,"DATA(%*d):%s",buf2);//TODO
-	a_modem_msg_add(&msg_remote,buf2);}
+	//sscanf(dump,"DATA(%*d):%s",buf2);//TODO
+	a_modem_msg_add(&msg_remote,strstr(dump,":")+1);}
 
 	return n;
 }
@@ -266,18 +274,20 @@ int a_modem_msg_send(const char*msg){
 	}
 	usleep(ONLINE_COMMAND);
 
-	a_modem_puts("+++");
+	a_modem_puts("+++\r");
 	usleep(COMMAND_DELAY);
 	return SUCCESS;
 }
 
 int a_modem_record(int duration) {
 // record waveform, store rx info as well.duration is in milliseconds
-	char buf[BUFSIZE], buf2[BUFSIZE];
+	char buf[BUFSIZE];
 	char logname[32];
-	int n;
-	/*record on*/
+	time_t bb_stamp;
+
 	modem.latest_rx_fname[0]=0;
+	time(&bb_stamp);
+	/*record on*/
 	a_modem_puts("\r");
 	a_modem_puts("record on\r");
 
@@ -285,21 +295,18 @@ int a_modem_record(int duration) {
 	usleep(duration * 1000); //TODO int overflow?
 
 	/*record off*/
-	a_modem_puts( "record off\r");
-	a_modem_puts("\r");
+	a_modem_puts( "record off\r\r");
 
 	/*get log name*/
 	if (a_modem_wait_info("log", 4*SERIAL_TIMEOUT, buf, BUFSIZE)==SUCCESS){
 	sscanf(buf,"%*s log file %s",logname);
-	printf("log name :%s\n",logname+4);
-	sprintf(buf2, "echo '%s' >> %s",logname+4,RX_PATH);
-	system(buf2);
-	system(RX_LOG);
+	printf("local log file : %s\n",logname+4);
 	strcpy(modem.latest_rx_fname,logname+4);
-
+	fprintf(modem.rx_p,"%s\n",logname+4);
+	fprintf(modem.rx_p,"bb : %s\n",ctime(&bb_stamp));
+	fflush(modem.rx_p);
 	return SUCCESS;
 	}else{
-	printf("debug : %s\n",buf);
 	fprintf(stderr, "A_modem, record msg (filename.log) missing\n");
 	return FAIL;
 	}
@@ -318,7 +325,8 @@ int a_modem_sync_time_gps() {
 	a_modem_clear_io_buffer();
 	// simple test
 	a_modem_puts("date\r");
-	a_modem_wait_info("2014", SERIAL_TIMEOUT, buf, BUFSIZE);
+	a_modem_wait_info(NULL, SERIAL_TIMEOUT, buf, BUFSIZE);
+	// dump
 	sprintf(buf2, "echo '%s' >> %s", buf,AMODEM_PATH);
 	system(buf2);//SYSTEM DUMP
 	a_modem_puts("date -store\r");
@@ -326,7 +334,7 @@ int a_modem_sync_time_gps() {
 	return SUCCESS;
 }
 
-int a_modem_sync_clock_gps() {
+int a_modem_sync_clock_gps(int sec) {
 	// sync modem clock source
 	a_modem_clear_io_buffer();
 	// Confirm clock source for the modem
@@ -335,15 +343,16 @@ int a_modem_sync_clock_gps() {
 		a_modem_puts("@SyncPPS=4\r");
 		printf("A_modem, syncpps source is not gps, reset the source...\n");
 		printf("sync..., this will take 30 seconds or more...\n");
-		sleep(30);
+		//sleep(30);
 	}
 	// confirm sync
 	a_modem_puts("sync\r");
 	printf("modem sync to pps signal...\n");
-	if (a_modem_is_clock_Sync(5, 6) == FAIL) {
+	if (a_modem_is_clock_Sync(sec) == FAIL) {
 		printf("A_modem, fail to sync\n");
 		return FAIL;
 	}
+	printf("local modem sync!\n");
 	return SUCCESS;
 }
 
@@ -377,23 +386,17 @@ int a_modem_sync_status() {
 	return FAIL;
 }
 
-int a_modem_is_clock_Sync(int samp_interval, int N_retry) {
-	//check if clock sync (regardless of clock source), samp_interval (sec)
+int a_modem_is_clock_Sync(int sec) {
+	//check if clock sync (regardless of clock source) until synchronized or timeout
 	int i;
-	// input check
-	a_modem_clear_io_buffer();
-	if ((samp_interval <= 0) | (N_retry <= 0)) {
-		printf("IS_SYNC, input error");
-		return FAIL;
-	}
 	// check every X sec
-	for (i = 0; i < N_retry; i++) {
+	for (i = 0; i < sec/(SERIAL_TIMEOUT/1000); i++) {
 		a_modem_puts("sync\r");
 		if (a_modem_wait_ack("synchronized", SERIAL_TIMEOUT)) {
 			modem.sync_state=SYNC;
 			return SUCCESS;
 		}
-		sleep(samp_interval);
+//		sleep(1);
 	}
 	modem.sync_state=NOT_SYNC;
 	printf("a_modem, sync time out\n");
@@ -421,6 +424,10 @@ int a_modem_wait_info(char *key_word, int timeout, char *info,
 			delay++;
 		} else {//got new msg
 			//new msg match
+			if (key_word==NULL){
+			if (info!=NULL) strncpy(info,msg.text[msg.i],info_size);
+			return SUCCESS;
+			}
 			if (strcasestr(msg.text[msg.i],key_word)!=NULL) {
 			if (info!=NULL) strncpy(info,msg.text[msg.i],info_size);
 			return SUCCESS;
@@ -496,27 +503,34 @@ int a_modem_upload_file(const char *fname){
 	char buf[BUFSIZE];
 	int ret;
 	int n_file;
-	memset(buf,0,BUFSIZE);
-	//check for existence
-	sprintf(buf,"ls -l /sd/%s\r",fname);
-	a_modem_puts(buf);
-	if (a_modem_wait_info("total",SERIAL_TIMEOUT,buf,BUFSIZE)==FAIL){
-		fprintf(stderr,"modem timeout\n");
-		return FAIL;
-	}
-	sscanf(buf,"total of %d file",&n_file);
-	if (n_file==0){
-		printf("file not exist\n");
-		return FAIL;
-	}
 	// copy
 	a_modem_clear_io_buffer();
 	sprintf(buf,"cp /sd/%s /ffs/%s\r",fname,fname);
 	a_modem_puts(buf);
-	sleep(1);
+	a_modem_wait_info(NULL,SERIAL_TIMEOUT,NULL,BUFSIZE);
+	if (a_modem_wait_info(NULL,COPY_TIMEOUT,buf,BUFSIZE)==FAIL){
+	fprintf(stderr,"cp no response\n");
+	return FAIL;
+	}
+	if (strstr(buf,"ok")==NULL){
+	printf("copy done \n");
+	}else{
+	fprintf(stderr,"cp error\n");
+	return FAIL;}
+	/*if (strcasestr(buf,"error")!=NULL){
+	fprintf(stderr,"cp error\n");
+	return FAIL;
+	}
+	if (strcasestr(buf,"ok")!=NULL){
+	fprintf(stdout,"cp done\n");
+	}else{
+	fprintf(stderr,"cp error\n");
+	return FAIL;
+	}*/
+	/*sleep(2);
 	buf[0]=0;
 	a_modem_gets(buf,BUFSIZE);
-	if (strcasestr(buf,"ok")!=NULL){
+	if (strcasestr(buf,"error")!=NULL){
 		fprintf(stderr,"fail to copy files in a modem\n");
 		return FAIL;
 	}
@@ -525,6 +539,7 @@ int a_modem_upload_file(const char *fname){
 		printf("copy file time out\r");
 		return FAIL;
 	}
+	*/
 	// issue ymodem send (sb)
 	sprintf(buf,"sb /ffs/%s\r",fname);
 	a_modem_puts(buf);
@@ -545,8 +560,16 @@ int a_modem_upload_file(const char *fname){
 	a_modem_puts(buf);
 	sprintf(buf,"rm /sd/%s\r",fname);
 	a_modem_puts(buf);
-	//move file
-	sprintf(buf,"mv ./%s /home/root/data",fname);
+	//move the file
+	sprintf(buf,"mv ./%s /home/root/data/.",fname);
 	system(buf);
 	return SUCCESS;
+}
+
+int a_modem_end(){
+fclose(modem.tx_p);
+fclose(modem.rx_p);
+a_modem_puts("record off\r");//make sure modem is not recording...
+a_modem_close();
+return SUCCESS;
 }
